# specification-module/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: |
    This is a daycare server.  You can find
    out more about Swagger at
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: 1.0.0
  title: Swagger Daycare
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/skyhigh-00b/DaycareAPI/1.0.0'
  - url: 'https://daycare.swagger.io/v2'
...
paths:
...
  /daycare:
    post:
      tags:
        - Daycare
      summary: Create new daycare
      description: This can be done by admin only
      operationId: createDayCare
      requestBody:
        description: Created daycare object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DayCareDto'
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayCareDto'
        '400':
          description: Operation failed
    put:
      tags:
        - Daycare
      summary: Edit daycare
      description: This can be done by admin or Daycare owner
      operationId: editDayCare
      requestBody:
        description: Created daycare object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DayCareDto'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayCareDto'
        '400':
          description: Operation failed
    get:
      tags:
        - Daycare
      summary: Get daycare list
      description: This can be done by any user
      operationId: listDayCares
      parameters:
        - name: keyword
          in: query
          description: Search keyword (searches name, description, etc.)
          required: false
          schema:
            type: string
            example: "Daycare name"
        - name: latitude
          in: query
          description: Latitude of the center point
          required: false
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
            example: 43.7615

        - name: longitude
          in: query
          description: Longitude of the center point
          required: false
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
            example: -79.4111

        - name: radius
          in: query
          description: Search radius in kilometers from the center point
          required: false
          schema:
            type: number
            format: float
            minimum: 0.1
            maximum: 100
            default: 10
            example: 25
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DayCareDto'
        '400':
          description: Operation failed
  /daycare/{id}:
    get:
      tags:
        - Daycare
      summary: Get daycare by id
      description: This can be done by any user
      operationId: getDayCare
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            example: 205
          required: true
          description: Numeric ID of the daycare to delete
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayCareDto'
        '400':
          description: Operation failed
    delete:
      tags:
        - Daycare
      summary: Delete daycare by id
      description: This can be done only by admin after receiving a written request from the daycare owner
      operationId: deleteDayCare
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            example: 206
          required: true
          description: Numeric ID of the daycare to delete
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DayCareDto'
        '400':
          description: Operation failed
...
components:
  schemas:
    UserTypes:
      type: string
      enum: [ admin, daycareOwner, daycareAssistant, parent, child ]
...
    OwnerDto:
      type: object
      description: Daycare owner
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        userType:
          $ref: '#/components/schemas/UserTypes'
    DayCareDto:
      type: object
      description: Day care DTO
      properties:
        name:
          type: string
        telephone:
          type: string
        owners:
          type: array
          items:
            $ref: '#/components/schemas/OwnerDto'
        address:
          $ref: '#/components/schemas/AddressDto'
    AddressDto:
      type: object
      properties:
        apartment:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
      description: Address model
...
  securitySchemes:
    daycare_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://daycare.swagger.io/oauth/dialog'
          scopes:
            'write:pets': modify pets in your account
            'read:pets': read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header